Updated jpaper to v2.2.1.
This version of jpaper changes the behavior of the "version" option so
that when both "version" and "anonymous" are set, the author field is
replaced by version information when available, with spacing tweaked
so that layout does not change. This is primarily useful for tablet
mode, for which there are no margins in which to display version
information otherwise.